blktrace: change size of action to 64 bits
In order to add the zoned commands to blktrace's actions, the storage size
needs to be increased to 64bits.

In order to keep the structure alignment 'action' and 'pid' need to be
swapped as well.

Therefore blktrace's version number also needs to be bumped.

+++ act_mask.c
@@ -42,7 +42,7 @@ int find_mask_map(char *string)
 	return -1;
 }
 
-int valid_act_opt(int x)
+unsigned long long valid_act_opt(unsigned long long x)
 {
-	return (1 <= x) && (x < (1 << BLK_TC_SHIFT));
+	return (1ull <= x) && (x < (1ull << BLK_TC_SHIFT));
 }
